refactor(camera): give the Auto camera its own file
PERSO.CPP is the translation unit that holds main() and the main loop, and
it has been collecting whatever lands near them: 2477 lines last October,
3999 now. The largest single contributor is the Auto camera, which has no
reason to sit next to the entry point.

Move the per-frame update, the camera trace, and the smoothing state they
carry between frames into FOLLOWCAM.CPP, next to the FOLLOWCAM_CFG.H and
FOLLOWCAM_MATH.H headers that were already split out.

It is not quite a pure relocation. The main loop was assigning into the
block's file-scope state in two places, so those assignments become
FollowCamResetToView() and FollowCamAdoptCurrentAngle(), named for the
moments that need them. That keeps the smoothing state private to the
module, which is the part that makes the file boundary worth having.
Everything else moves verbatim.

Verified with a camtrace A/B: same save, same fixed timestep, same nudge
timeline, 77 traced frames identical before and after. The seven camera
fixtures in tests/automation pass, and host_quick is unchanged.
